Add Ubuntu 26.04

+++ install.sh
@@ -19,8 +19,23 @@ if [[ $collections =~ desktop ]]; then
   echo "Installation in progress, please wait" > /etc/nologin
 fi
 
-# Get the latest package info and install any updates
 if [[ $dist == ubuntu ]]; then
+  # Install Firefox using DEB package rather than snap to avoid timeout issues
+  install -d -m 0755 /etc/apt/keyrings
+  wget -q https://packages.mozilla.org/apt/repo-signing-key.gpg -O- | tee /etc/apt/keyrings/packages.mozilla.org.asc > /dev/null || error
+  tee /etc/apt/sources.list.d/mozilla.sources > /dev/null << EOF
+Types: deb
+URIs: https://packages.mozilla.org/apt
+Suites: mozilla
+Components: main
+Signed-By: /etc/apt/keyrings/packages.mozilla.org.asc
+EOF
+  tee /etc/apt/preferences.d/mozilla > /dev/null << EOF
+Package: *
+Pin: origin packages.mozilla.org
+Pin-Priority: 1000
+EOF
+  # Get the latest package info and install any updates
   export DEBIAN_FRONTEND=noninteractive  # Disable user interaction
   apt-get -yq update || error
   apt-get -yq -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" upgrade || error
@@ -32,16 +47,13 @@ if [[ $dist == ubuntu ]]; then
 fi
 
 for collection in $collections; do
-  echo $(date +"%Y-%m-%dT%H%M") - Installing $collection collection ...
-  dos2unix -n /vagrant/install-$collection.sh /tmp/install-$collection.sh
-  . /tmp/install-$collection.sh
-  rm /tmp/install-$collection.sh
+  if [[ ! $collection == skip-shutdown ]]; then
+    echo $(date +"%Y-%m-%dT%H%M") - Installing $collection collection ...
+    dos2unix -n /vagrant/install-$collection.sh /tmp/install-$collection.sh
+    . /tmp/install-$collection.sh
+    rm /tmp/install-$collection.sh
+  fi
 done
-
-# Remove python-gi on Ubuntu since it breaks rosie go (not needed unless using GNOME keyring)
-if [[ $dist == ubuntu ]]; then
-  apt-get remove -q -y --auto-remove --purge python-gi || error
-fi
 
 set +x
 echo Finished provisioning at $(date +"%Y-%m-%dT%H%M") \(started at $STARTDATE\)
